make class to make new accounts and repair some exceptions in FileReader and some future to class admin

+++ lib_h/Input.h
@@ -9,7 +9,7 @@
 
 class Input {
         public:
-           ll getInt();
+           long long getInt();
            double getDouble();
            bool getBool();
            char getChar();
